+++ lib/ui/events2/Event2CreatePanel.dart
@@ -1540,7 +1540,7 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
   }
 
   void _onMonthDayChanged(int? day) {
-    Analytics().logSelect(target: "Recurrence Every month: $day");
+    Analytics().logSelect(target: "Recurrence Every month day: $day");
     Event2CreatePanel.hideKeyboard(context);
     setStateIfMounted(() {
       _recurrenceRepeatDay = day;
@@ -2459,12 +2459,26 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
       _creatingEvent = true;
     });
 
+    List<_RecurringDatesPair>? recurringDates = _buildRecurringDatesPairs();
+    DateTime? eventStartDate, eventEndDate;
+    if (CollectionUtils.isNotEmpty(recurringDates)) {
+      eventStartDate = recurringDates!.first.startDateTimeUtc;
+      eventEndDate = recurringDates.last.endDateTimeUtc;
+    }
+
     dynamic result;
-    Event2 event = _createEventFromData();
+    // Explicitly set the start date to be the first and end date to be the last - #4599
+    Event2 event = _createEventFromData(recurringStartDateUtc: eventStartDate, recurringEndDateUtc: eventEndDate);
 
     //TBD: DD - tmp code for debug purposes
     // //////// remove start
-    // await _createRecurringEventsFrom(mainEvent: event);
+    // print('TTTTTTTTTTTT: main: start: ${event.startTimeUtc}, end: ${event.endTimeUtc}');
+    // if (CollectionUtils.isNotEmpty(recurringDates)) {
+    //   for (_RecurringDatesPair pair in recurringDates!) {
+    //     print('TTTTTTTTTTTT: recurring: start: ${pair.startDateTimeUtc}, end: ${pair.endDateTimeUtc}');
+    //   }
+    // }
+    //
     // setState(() {
     //   _creatingEvent = false;
     // });
@@ -2488,8 +2502,8 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
       if (result is Event2) {
         Survey? survey = widget.survey;
         if (widget.isCreate) {
-          if (_shouldCreateRecurringEvents) {
-            await _createRecurringEventsFrom(mainEvent: result);
+          if (_shouldCreateRecurringEvents && CollectionUtils.isNotEmpty(recurringDates)) {
+            await _createRecurringEventsFrom(mainEvent: result, recurringDates: recurringDates!);
           }
           if (_survey != null) {
             bool? success = await Surveys().createEvent2Survey(_survey!, result);
@@ -2678,6 +2692,7 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
     ((_registrationDetails?.type != Event2RegistrationType.external) || (_registrationDetails?.externalLink?.isNotEmpty ?? false)) &&
     ((_registrationDetails?.type != Event2RegistrationType.internal) || ((_registrationDetails?.eventCapacity ?? 0) > 0)) &&
     (!_hasSurvey || _hasAttendanceDetails)
+      //TBD: DD - implement for recurrence
   );
 
   void _updateErrorMap() {
@@ -2778,7 +2793,7 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
     }
   }
 
-  Event2 _createEventFromData() {
+  Event2 _createEventFromData({DateTime? recurringStartDateUtc, DateTime? recurringEndDateUtc}) {
     List<String>? groupIds = _eventGroups?.map((group) => group.id!).toList();
     Event2AuthorizationContext? authorizationContext;
     Event2Context? event2Context;
@@ -2807,6 +2822,9 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
       grouping = widget.event?.grouping;
     }
 
+    DateTime? eventStartDateUtc = (widget.isCreate && _shouldCreateRecurringEvents && (recurringStartDateUtc != null)) ? recurringStartDateUtc : _startDateTimeUtc;
+    DateTime? eventEndDateUtc = (widget.isCreate && _shouldCreateRecurringEvents && (recurringEndDateUtc != null)) ? recurringEndDateUtc : _endDateTimeUtc;
+
     return Event2(
       id: widget.event?.id,
 
@@ -2817,8 +2835,8 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
       eventUrl: Event2CreatePanel.textFieldValue(_websiteController),
 
       timezone: _timeZone.name,
-      startTimeUtc: _startDateTimeUtc,
-      endTimeUtc: (widget.isCreate && _shouldCreateRecurringEvents) ? _recurrenceEndDateTimeUtc : _endDateTimeUtc,
+      startTimeUtc: eventStartDateUtc,
+      endTimeUtc: eventEndDateUtc,
       allDay: _allDay,
 
       eventType: _eventType,
@@ -2847,9 +2865,256 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
     );
   }
 
-  Future<void> _createRecurringEventsFrom({required Event2 mainEvent}) async {
+  List<_RecurringDatesPair>? _buildRecurringDatesPairs() {
+    if (!widget.isCreate || !_shouldCreateRecurringEvents) {
+      return null;
+    }
+    List<_RecurringDatesPair>? recurringDates;
+    switch (_recurrenceRepeatType) {
+      case _RecurrenceRepeatType.weekly:
+        recurringDates = _buildWeeklyRecurringDates();
+        break;
+      case _RecurrenceRepeatType.monthly:
+        recurringDates = _buildMonthlyRecurringDates();
+        break;
+      default:
+        break;
+    }
+    return recurringDates;
+  }
+
+  List<_RecurringDatesPair>? _buildWeeklyRecurringDates() {
+    List<int>? recurrenceWeekDaysIndexes = _recurrenceWeekDays?.map((day) => day.index).toList();
+    recurrenceWeekDaysIndexes?.sort();
+    DateTime recurringEndDateTimeUtc = _recurrenceEndDateTimeUtc!;
+    List<_RecurringDatesPair> pairs = <_RecurringDatesPair>[];
+    DateTime nextStartDateUtc = _startDateTimeUtc!;
+    DateTime nextEndDateUtc = _endDateTimeUtc!;
+    while (nextStartDateUtc.isBefore(recurringEndDateTimeUtc)) {
+      if (recurrenceWeekDaysIndexes?.contains(nextStartDateUtc.weekday - 1) ?? false) {
+        pairs.add(_RecurringDatesPair(startDateTimeUtc: nextStartDateUtc, endDateTimeUtc: nextEndDateUtc));
+      }
+      int daysToAdd = (nextStartDateUtc.weekday == 7) ? (1 + (_weeklyRepeatPeriod! - 1) * 7) : 1;
+      nextStartDateUtc = nextStartDateUtc.add(Duration(days: daysToAdd));
+      nextEndDateUtc = nextEndDateUtc.add(Duration(days: daysToAdd));
+    }
+    return pairs;
+  }
+
+  List<_RecurringDatesPair>? _buildMonthlyRecurringDates() {
+    List<_RecurringDatesPair>? pairs;
+    switch (_recurrenceRepeatMonthlyType) {
+      case _RecurrenceRepeatMonthlyType.daily:
+        pairs = _buildMonthlyRecurringDatesByOrdinalDay();
+        break;
+      case _RecurrenceRepeatMonthlyType.weekly:
+        pairs = _buildMonthlyRecurringDatesByWeekDay();
+        break;
+      default:
+        break;
+    }
+    return pairs;
+  }
+
+  List<_RecurringDatesPair>? _buildMonthlyRecurringDatesByOrdinalDay() {
+    DateTime recurringEndDateTimeUtc = _recurrenceEndDateTimeUtc!;
+    List<_RecurringDatesPair> pairs = <_RecurringDatesPair>[];
+    DateTime nextStartDateUtc = _startDateTimeUtc!;
+    DateTime nextEndDateUtc = _endDateTimeUtc!;
+    while (nextStartDateUtc.isBefore(recurringEndDateTimeUtc)) {
+      if ((_recurrenceRepeatDay == 0) || (_recurrenceRepeatDay == nextStartDateUtc.day)) {
+        pairs.add(_RecurringDatesPair(startDateTimeUtc: nextStartDateUtc, endDateTimeUtc: nextEndDateUtc));
+      }
+
+      late int daysDiff;
+      if (_recurrenceRepeatDay == 0) {
+        daysDiff = 1;
+      } else {
+        if (nextStartDateUtc.day < _recurrenceRepeatDay!) {
+          daysDiff = (_recurrenceRepeatDay! - nextStartDateUtc.day);
+        } else {
+          DateTime nextDate = DateTime.utc(
+              nextStartDateUtc.year,
+              (nextStartDateUtc.month + _monthlyRepeatPeriod!),
+              _recurrenceRepeatDay!,
+              nextStartDateUtc.hour,
+              nextStartDateUtc.minute,
+              nextStartDateUtc.second,
+              nextStartDateUtc.millisecond,
+              nextStartDateUtc.microsecond);
+          daysDiff = nextDate.difference(nextStartDateUtc).inDays;
+        }
+      }
+
+      Duration duration = Duration(days: daysDiff);
+      nextStartDateUtc = nextStartDateUtc.add(duration);
+      nextEndDateUtc = nextEndDateUtc.add(duration);
+    }
+    return pairs;
+  }
+
+  List<_RecurringDatesPair>? _buildMonthlyRecurringDatesByWeekDay() {
+    DateTime recurringEndDateTimeUtc = _recurrenceEndDateTimeUtc!;
+    List<_RecurringDatesPair> pairs = <_RecurringDatesPair>[];
+    DateTime nextStartDateUtc = _startDateTimeUtc!;
+    DateTime nextEndDateUtc = _endDateTimeUtc!;
+    int? nThDayOfMonth = _nThDayOfMonth;
+    DateTime? desiredDateTime = _getInitialRecurringDesiredDay(nextStartDateUtc: nextStartDateUtc, nThDayOfMonth: nThDayOfMonth);
+    if (desiredDateTime != null) {
+      while (nextStartDateUtc.isBefore(recurringEndDateTimeUtc)) {
+        if (nextStartDateUtc.day == desiredDateTime!.day) {
+          pairs.add(_RecurringDatesPair(startDateTimeUtc: nextStartDateUtc, endDateTimeUtc: nextEndDateUtc));
+        } else if (nextStartDateUtc.day < desiredDateTime.day) {
+          int daysDiff = desiredDateTime.difference(nextStartDateUtc).inDays;
+          nextStartDateUtc = nextStartDateUtc.add(Duration(days: daysDiff));
+          nextEndDateUtc = nextEndDateUtc.add(Duration(days: daysDiff));
+          pairs.add(_RecurringDatesPair(startDateTimeUtc: nextStartDateUtc, endDateTimeUtc: nextEndDateUtc));
+        }
+        desiredDateTime = _getNextRecurringDesiredDay(nextStartDateUtc: nextStartDateUtc, nThDayOfMonth: nThDayOfMonth);
+        int daysDiffToNext = desiredDateTime!.difference(nextStartDateUtc).inDays;
+        Duration duration = Duration(days: daysDiffToNext);
+        nextStartDateUtc = nextStartDateUtc.add(duration);
+        nextEndDateUtc = nextEndDateUtc.add(duration);
+      }
+    }
+    return pairs;
+  }
+
+  DateTime? _getInitialRecurringDesiredDay({required DateTime nextStartDateUtc, int? nThDayOfMonth}) {
+    DateTime? dateTime;
+    if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.day) {
+      int month = (nThDayOfMonth != null) ? nextStartDateUtc.month : (nextStartDateUtc.month + 1);
+      int day = (nThDayOfMonth != null) ? nThDayOfMonth : 0;
+      dateTime = DateTime.utc(nextStartDateUtc.year, month, day, nextStartDateUtc.hour, nextStartDateUtc.minute, nextStartDateUtc.second,
+          nextStartDateUtc.millisecond, nextStartDateUtc.microsecond);
+    } else if (_isSelectedNamedWeekDay) {
+      dateTime = (nThDayOfMonth != null)
+          ? _getNthNamedWeekDay(dateTimeUtc: nextStartDateUtc, n: nThDayOfMonth, nextMonth: false, targetWeekDayNumber: _monthWeekDayNumber)
+          : _getLastNamedWeekDay(dateTimeUtc: nextStartDateUtc, nextMonth: false, targetWeekDayNumber: _monthWeekDayNumber);
+    } else if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.weekday) {
+      //TBD: DD - implement
+    } else if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.weekend_day) {
+      //TBD: DD - implement
+    }
+    return dateTime;
+  }
+
+  DateTime? _getNextRecurringDesiredDay({required DateTime nextStartDateUtc, int? nThDayOfMonth}) {
+    DateTime? dateTime;
+    if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.day) {
+      int month = ((nThDayOfMonth != null) ? nextStartDateUtc.month : (nextStartDateUtc.month + 1)) + _monthlyRepeatPeriod!;
+      int day = (nThDayOfMonth != null) ? nThDayOfMonth : 0;
+      dateTime = DateTime.utc(nextStartDateUtc.year, month, day, nextStartDateUtc.hour, nextStartDateUtc.minute, nextStartDateUtc.second,
+          nextStartDateUtc.millisecond, nextStartDateUtc.microsecond);
+    } else if (_isSelectedNamedWeekDay) {
+      dateTime = (nThDayOfMonth != null)
+          ? _getNthNamedWeekDay(
+              dateTimeUtc: nextStartDateUtc,
+              n: nThDayOfMonth,
+              nextMonth: true,
+              targetWeekDayNumber: _monthWeekDayNumber,
+              monthsAhead: _monthlyRepeatPeriod)
+          : _getLastNamedWeekDay(
+              dateTimeUtc: nextStartDateUtc, nextMonth: true, targetWeekDayNumber: _monthWeekDayNumber, monthsAhead: _monthlyRepeatPeriod);
+    } else if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.weekday) {
+      //TBD: DD - implement
+    } else if (_recurrenceMonthWeekDay == _RecurrenceMonthWeekDay.weekend_day) {
+      //TBD: DD - implement
+    }
+    return dateTime;
+  }
+
+  bool get _isSelectedNamedWeekDay {
+    return (_recurrenceMonthWeekDay != null) &&
+        ![_RecurrenceMonthWeekDay.day, _RecurrenceMonthWeekDay.weekday, _RecurrenceMonthWeekDay.weekend_day]
+            .contains(_recurrenceMonthWeekDay);
+  }
+
+  DateTime? _getNthNamedWeekDay(
+      {required DateTime dateTimeUtc, required int n, bool nextMonth = false, int? targetWeekDayNumber, int? monthsAhead}) {
+    if ((targetWeekDayNumber == null) || (targetWeekDayNumber < 1) || (targetWeekDayNumber > 7)) {
+      return null;
+    }
+    int month = nextMonth ? (dateTimeUtc.month + (monthsAhead ?? 1)) : dateTimeUtc.month;
+    DateTime firstDayOfMonth = DateTime.utc(dateTimeUtc.year, month, 1, dateTimeUtc.hour, dateTimeUtc.minute, dateTimeUtc.second,
+        dateTimeUtc.millisecond, dateTimeUtc.microsecond);
+    int firstDayOfWeek = firstDayOfMonth.weekday;
+    int daysUntilTarget = (targetWeekDayNumber - firstDayOfWeek + 7) % 7;
+    DateTime targetDay = firstDayOfMonth.add(Duration(days: daysUntilTarget));
+
+    if ((n > 1) && (n <= 4)) {
+      targetDay = targetDay.add(Duration(days: (n - 1) * 7));
+    }
+    if (targetDay.month != month) {
+      return null;
+    }
+
+    return targetDay;
+  }
+
+  DateTime? _getLastNamedWeekDay({required DateTime dateTimeUtc, bool nextMonth = false, int? targetWeekDayNumber, int? monthsAhead}) {
+    if ((targetWeekDayNumber == null) || (targetWeekDayNumber < 1) || (targetWeekDayNumber > 7)) {
+      return null;
+    }
+    int month = (nextMonth ? (dateTimeUtc.month + (monthsAhead ?? 1)) : dateTimeUtc.month) + 1;
+    DateTime lastDayOfMonth = DateTime.utc(dateTimeUtc.year, month, 0, dateTimeUtc.hour, dateTimeUtc.minute, dateTimeUtc.second,
+        dateTimeUtc.millisecond, dateTimeUtc.microsecond);
+    int lastDayOfWeek = lastDayOfMonth.weekday;
+    int daysUntilTargetDay = (lastDayOfWeek - targetWeekDayNumber + 7) % 7;
+    DateTime targetDay = lastDayOfMonth.subtract(Duration(days: daysUntilTargetDay));
+    return targetDay;
+  }
+
+  int? get _monthWeekDayNumber {
+    switch (_recurrenceMonthWeekDay) {
+      case _RecurrenceMonthWeekDay.monday:
+        return 1;
+      case _RecurrenceMonthWeekDay.tuesday:
+        return 2;
+      case _RecurrenceMonthWeekDay.wednesday:
+        return 3;
+      case _RecurrenceMonthWeekDay.thursday:
+        return 4;
+      case _RecurrenceMonthWeekDay.friday:
+        return 5;
+      case _RecurrenceMonthWeekDay.saturday:
+        return 6;
+      case _RecurrenceMonthWeekDay.sunday:
+        return 7;
+      default:
+        return null;
+    }
+  }
+
+  int? get _nThDayOfMonth {
+    switch (_recurrenceOrdinalNumber) {
+      case _RecurrenceOrdinalNumber.first:
+        return 1;
+      case _RecurrenceOrdinalNumber.second:
+        return 2;
+      case _RecurrenceOrdinalNumber.third:
+        return 3;
+      case _RecurrenceOrdinalNumber.fourth:
+        return 4;
+      default:
+        return null;
+    }
+  }
+
+  List<Event2>? _buildRecurringEventsFrom({required Event2 mainEvent, required List<_RecurringDatesPair> dates}) {
+    if (CollectionUtils.isEmpty(dates)) {
+      return null;
+    }
+    List<Event2> events = <Event2>[];
+    for (_RecurringDatesPair pair in dates) {
+      events.add(mainEvent.toRecurringEvent(startDateTimeUtc: pair.startDateTimeUtc, endDateTimeUtc: pair.endDateTimeUtc));
+    }
+    return events;
+  }
+
+  Future<void> _createRecurringEventsFrom({required Event2 mainEvent, required List<_RecurringDatesPair> recurringDates}) async {
     // Create each event separately until we have a backend API for that
-    List<Event2>? recurringEvents = _buildRecurringEventsFrom(mainEvent: mainEvent);
+    List<Event2>? recurringEvents = _buildRecurringEventsFrom(mainEvent: mainEvent, dates: recurringDates);
     if (CollectionUtils.isNotEmpty(recurringEvents)) {
       for (Event2 recurringEvent in recurringEvents!) {
         dynamic recurringResult = await Events2().createEvent(recurringEvent);
@@ -2866,61 +3131,6 @@ class _Event2CreatePanelState extends State<Event2CreatePanel> {
         }
       }
     }
-  }
-
-  List<Event2>? _buildRecurringEventsFrom({required Event2 mainEvent}) {
-    List<Event2>? recurringEvents;
-    switch (_recurrenceRepeatType) {
-      case _RecurrenceRepeatType.weekly:
-        recurringEvents = _buildWeeklyRecurringEvents(mainEvent: mainEvent);
-        break;
-      case _RecurrenceRepeatType.monthly:
-        recurringEvents = _buildMonthlyRecurringEvents(mainEvent: mainEvent);
-        break;
-      default:
-        break;
-    }
-    return recurringEvents;
-  }
-
-  List<Event2>? _buildWeeklyRecurringEvents({required Event2 mainEvent}) {
-    List<int>? recurrenceWeekDaysIndexes = _recurrenceWeekDays?.map((day) => day.index).toList();
-    recurrenceWeekDaysIndexes?.sort();
-    DateTime mainStartDateTimeUtc = _startDateTimeUtc!;
-    DateTime mainEndDateTimeUtc = _endDateTimeUtc!;
-    DateTime recurringEndDateTimeUtc = _recurrenceEndDateTimeUtc!;
-    List<DateTime> startDateTimesUtc = <DateTime>[];
-    List<DateTime> endDateTimesUtc = <DateTime>[];
-    DateTime nextStartDateUtc = mainStartDateTimeUtc.add(Duration(days: 1));
-    DateTime nextEndDateUtc = mainEndDateTimeUtc.add(Duration(days: 1));
-    while (nextStartDateUtc.isBefore(recurringEndDateTimeUtc)) {
-      if (recurrenceWeekDaysIndexes?.contains(nextStartDateUtc.weekday - 1) ?? false) {
-        startDateTimesUtc.add(nextStartDateUtc);
-        endDateTimesUtc.add(nextEndDateUtc);
-      }
-      int daysToAdd = (nextStartDateUtc.weekday == 7) ? (1 + (_weeklyRepeatPeriod! - 1) * 7) : 1;
-      nextStartDateUtc = nextStartDateUtc.add(Duration(days: daysToAdd));
-      nextEndDateUtc = nextEndDateUtc.add(Duration(days: daysToAdd));
-    }
-
-    List<Event2>? events;
-    if (CollectionUtils.isNotEmpty(startDateTimesUtc)) {
-      events = <Event2>[];
-      for (int i = 0; i < startDateTimesUtc.length; i++) {
-        DateTime start = startDateTimesUtc[i];
-        DateTime end = endDateTimesUtc[i];
-        Event2? subEvent = mainEvent.toRecurringEvent(startDateTimeUtc: start, endDateTimeUtc: end);
-        if (subEvent != null) {
-          events.add(subEvent);
-        }
-      }
-    }
-    return events;
-  }
-
-  List<Event2>? _buildMonthlyRecurringEvents({required Event2 mainEvent}) {
-    //TBD: DD - implement
-    return null;
   }
 }
 
@@ -3007,4 +3217,11 @@ String _recurrenceMonthWeekDayToDisplayString(_RecurrenceMonthWeekDay? value) {
     case _RecurrenceMonthWeekDay.weekend_day: return Localization().getStringEx('panel.event2.create.recurrence.month.weekday.weekend_day.label', 'Weekend Day');
     default: return '-----';
   }
+}
+
+class _RecurringDatesPair {
+  final DateTime startDateTimeUtc;
+  final DateTime endDateTimeUtc;
+
+  _RecurringDatesPair({required this.startDateTimeUtc, required this.endDateTimeUtc});
 }